// src/modules/spouseInformation/spouseInformation.interface.ts
export type SpouseInformationField =
  | "full_name"
  | "contact_no"
  | "father_name"
  | "father_contact_no"
  | "mother_name"
  | "mother_contact_no"
  | "occupation"
  | "job_title"
  | "monthly_income";

  export type SpouseInformationValue = string | null;

export type SpouseInformationDeleteField =
  | "contact_no"
  | "father_name"
  | "father_contact_no"
  | "mother_name"
  | "mother_contact_no"
  | "occupation"
  | "job_title"
  | "monthly_income";

export type SpouseInformationDeleteValue = string | null;